% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCityMap.R
\name{viewCityMap}
\alias{viewCityMap}
\title{viewCityMap}
\usage{
viewCityMap(city_name, state_abb)
}
\arguments{
\item{city_name}{The name of the city being mapped}

\item{state_abb}{The abbreviation of the state that the city is in. (e.g.
New York = "NY")}
}
\description{
A function that allows users to create a map of a city in the
United States and its surrounding area using Stamen Maps.
}
\details{
Maps are generated by using package ggmap to pull map data from
Stamen Maps

D. Kahle and H. Wickham. ggmap: Spatial Visualization with ggplot2. The R
Journal, 5(1), 144-161.
URL \url{http://journal.r-project.org/archive/2013-1/kahle-wickham.pdf}

US city data comes from the simplemaps United States Cities Database
\url{https://simplemaps.com/data/us-cities}
}
